requirements.txt
wheel
scikit-image
numpy
pillow
matplotlib
qrcode
gpiozero
picamera2
